Reorganize project files structure, add strong password validation
