+++ src/users/dto/user.dto.ts
@@ -1,5 +1,11 @@
 import { ApiProperty, OmitType } from "@nestjs/swagger";
-import { IsEmail, IsNotEmpty, IsString, MinLength } from "class-validator";
+import {
+  IsEmail,
+  IsNotEmpty,
+  IsString,
+  IsStrongPassword,
+  MinLength,
+} from "class-validator";
 
 export class CreateUserDto {
   @ApiProperty({
@@ -32,7 +38,13 @@ export class CreateUserDto {
   })
   @IsNotEmpty()
   @IsString()
-  @MinLength(6)
+  @IsStrongPassword({
+    minLength: 8,
+    minLowercase: 1,
+    minNumbers: 1,
+    minUppercase: 1,
+    minSymbols: 1,
+  })
   password: string;
 }
