--- src/users/users.controller.ts
@@ -1,7 +1,0 @@
-import { Controller } from "@nestjs/common";
-import { UsersService } from "./users.service";
-
-@Controller("users")
-export class UsersController {
-  constructor(private usersService: UsersService) {}
-}
