+++ src/users/users.module.ts
@@ -1,5 +1,4 @@
 import { Module } from "@nestjs/common";
-import { UsersController } from "./users.controller";
 import { UsersService } from "./users.service";
 import { MongooseModule } from "@nestjs/mongoose";
 import { User, UserSchema } from "./schemas/user.schema";
@@ -8,7 +7,6 @@ import { User, UserSchema } from "./schemas/user.schema";
   imports: [
     MongooseModule.forFeature([{ name: User.name, schema: UserSchema }]),
   ],
-  controllers: [UsersController],
   providers: [UsersService],
   exports: [
     UsersService,
